Deleted users logout from the system
- Added redirect to log out action, then already deleted user tries to
take any action in his profile.

+++ Web/Controllers/Profile/ProfileController.cs
@@ -17,6 +17,9 @@ namespace Template.Controllers.Profile
         [HttpGet]
         public ActionResult Edit()
         {
+            if (!Service.AccountExists(HttpContext.User.Identity.Name))
+                return LogOut();
+
             return View(Service.GetView(HttpContext.User.Identity.Name));
         }
 
@@ -24,6 +27,9 @@ namespace Template.Controllers.Profile
         [ValidateAntiForgeryToken]
         public ActionResult Edit(ProfileView profile)
         {
+            if (!Service.AccountExists(HttpContext.User.Identity.Name))
+                return LogOut();
+
             if (Service.CanEdit(profile))
                 Service.Edit(profile);
 
@@ -33,6 +39,9 @@ namespace Template.Controllers.Profile
         [HttpGet]
         public ActionResult Delete()
         {
+            if (!Service.AccountExists(HttpContext.User.Identity.Name))
+                return LogOut();
+
             ProfileView profile = Service.GetView(HttpContext.User.Identity.Name);
             Service.AddDeleteDisclaimerMessage();
             profile.Username = String.Empty;
@@ -45,10 +54,18 @@ namespace Template.Controllers.Profile
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(ProfileView profile)
         {
+            if (!Service.AccountExists(HttpContext.User.Identity.Name))
+                return LogOut();
+
             if (!Service.CanDelete(profile))
                 return View();
 
             Service.Delete(HttpContext.User.Identity.Name);
+            return LogOut();
+        }
+
+        private RedirectToRouteResult LogOut()
+        {
             return RedirectToAction("Logout", "Account");
         }
     }
